+++ R/resmush-package.R
@@ -1,3 +1,14 @@
+#' @seealso
+#' - [resmush_file()] optimizes local image files.
+#' - [resmush_dir()] optimizes image files in directories.
+#' - [resmush_url()] optimizes online image files.
+#' - The [package website](https://dieghernan.github.io/resmush/) provides
+#'   articles and reference documentation.
+#' - The [source code](https://github.com/dieghernan/resmush) is available on
+#'   GitHub.
+#' - [Report bugs](https://github.com/dieghernan/resmush/issues) in the issue
+#'   tracker.
+#'
 #' @keywords internal
 #' @encoding UTF-8
 "_PACKAGE"